Add options field to struct substitute_{function,import}_hook.
I will probably need to add an option to deal with GCC 5.2's new
optimization - to not save caller-saved registers if the called function
is known not to use them, which would break with hooked functions.
Don't want to emit the required stub by default because it will likely
be slow.  I need to verify that Clang doesn't currently have such an
optimization...

(ABI break - need to avoid this once it gets widely distributed Soon,
obviously)